# Fix infinite recursion indexing poorly constructed source JARs.