Indexing source JARs can result in infinite recursion when indexing a poorly constructed source JAR causing high CPU, the metals JVM heap to grow until exhausting available memory and workspace indexing to never complete.

Happy to make any style or refactoring changes. Existing tests pass. Benchmark is comparable - see below. Adding a specific test would really require committing `openhtmltopdf-core-0.0.1-RC19-sources.jar` to the source repository, which can be done if necessary.

**Reproduction**
1. VSCode 1.44.2
2. OpenJDK 11
3. Metals 0.9.0+47-6e4a2f22-SNAPSHOT, 0.9.0, 0.8.4
4. Checkout Metals master branch with VSCode and Metals extension installed.
5. Import build with Metals

**Problem**
The JAR `openhtmltopdf-core-0.0.1-RC19-sources.jar` is poorly constructed containing a root directory `/` in the root of the zip file causing `CommonMtagsEnrichments.listRecursive` to perpetually walk this directory and those below it - see below. This PR replaces the use of `CommonMtagsEnrichments.listRecursive` with a `nio.file.Files.walkFileTree` using a `nio.file.FileVisitor` that visits each path in the zip file at-most-once. It may be worth considering the use of `CommonMtagsEnrichments.listRecursive` in other areas of the code base and whether or not an at-most-once implementation is preferable.

**Benchmark**
Before:
```
[info] Benchmark                 Mode  Cnt  Score   Error  Units
[info] MetalsBench.indexSources    ss    5  3.454 ± 2.440   s/op
```
After:
```
[info] Benchmark                 Mode  Cnt  Score   Error  Units
[info] MetalsBench.indexSources    ss    5  3.935 ± 1.111   s/op
```

**Poorly Constructed Source JAR**
```
unzip -l ~/.cache/coursier/v1/https/repo1.maven.org/maven2/com/openhtmltopdf/openhtmltopdf-core/0.0.1-RC19/openhtmltopdf-core-0.0.1-RC19-sources.jar
Archive:  
~/.cache/coursier/v1/https/repo1.maven.org/maven2/com/openhtmltopdf/openhtmltopdf-core/0.0.1-RC19/openhtmltopdf-core-0.0.1-RC19-sources.jar
  Length      Date    Time    Name
---------  ---------- -----   ----
        ...
        0  2019-02-09 18:14   /
        0  2018-09-28 15:41   /Users/
        0  2019-03-02 18:54   /Users/daniel/
        0  2019-03-11 18:07   /Users/daniel/Documents/
        0  2017-11-07 17:25   /Users/daniel/Documents/workspace/
        0  2019-03-18 20:36   /Users/daniel/Documents/workspace/flyingsaucer/
        0  2019-03-18 20:36   /Users/daniel/Documents/workspace/flyingsaucer/target/
        0  2019-03-18 20:37   /Users/daniel/Documents/workspace/flyingsaucer/target/checkout/
        0  2019-03-18 20:38   /Users/daniel/Documents/workspace/flyingsaucer/target/checkout/openhtmltopdf-core/
        0  2019-03-18 20:38   /Users/daniel/Documents/workspace/flyingsaucer/target/checkout/openhtmltopdf-core/target/
        0  2019-03-18 20:38   /Users/daniel/Documents/workspace/flyingsaucer/target/checkout/openhtmltopdf-core/target/classes/
        0  2019-03-18 20:38   /Users/daniel/Documents/workspace/flyingsaucer/target/checkout/openhtmltopdf-core/target/classes/META-INF/
        ...
```

**Stacktrace**
```
Thread 0x601dbbb00
  at java.util.Arrays.hashCode([B)I (Arrays.java:4569)
  at jdk.nio.zipfs.ZipFileSystem$IndexNode.name([B)V (ZipFileSystem.java:1825)
  at jdk.nio.zipfs.ZipFileSystem$IndexNode.as([B)Ljdk/nio/zipfs/ZipFileSystem$IndexNode; (ZipFileSystem.java:1829)
  at jdk.nio.zipfs.ZipFileSystem$IndexNode.keyOf([B)Ljdk/nio/zipfs/ZipFileSystem$IndexNode; (ZipFileSystem.java:1820)
  at jdk.nio.zipfs.ZipFileSystem.getInode([B)Ljdk/nio/zipfs/ZipFileSystem$IndexNode; (ZipFileSystem.java:1325)
  at jdk.nio.zipfs.ZipFileSystem.isDirectory([B)Z (ZipFileSystem.java:384)
  at jdk.nio.zipfs.ZipDirectoryStream.<init>(Ljdk/nio/zipfs/ZipPath;Ljava/nio/file/DirectoryStream$Filter;)V (ZipDirectoryStream.java:57)
  at jdk.nio.zipfs.ZipPath.newDirectoryStream(Ljava/nio/file/DirectoryStream$Filter;)Ljava/nio/file/DirectoryStream; (ZipPath.java:714)
  at jdk.nio.zipfs.ZipFileSystemProvider.newDirectoryStream(Ljava/nio/file/Path;Ljava/nio/file/DirectoryStream$Filter;)Ljava/nio/file/DirectoryStream; (ZipFileSystemProvider.java:260)
  at java.nio.file.Files.newDirectoryStream(Ljava/nio/file/Path;)Ljava/nio/file/DirectoryStream; (Files.java:471)
  at java.nio.file.FileTreeWalker.visit(Ljava/nio/file/Path;ZZ)Ljava/nio/file/FileTreeWalker$Event; (FileTreeWalker.java:300)
  at java.nio.file.FileTreeWalker.next()Ljava/nio/file/FileTreeWalker$Event; (FileTreeWalker.java:373)
  at java.nio.file.FileTreeIterator.fetchNextIfNeeded()V (FileTreeIterator.java:83)
  at java.nio.file.FileTreeIterator.hasNext()Z (FileTreeIterator.java:103)
  at java.util.Spliterators$IteratorSpliterator.tryAdvance(Ljava/util/function/Consumer;)Z (Spliterators.java:1811)
  at java.util.stream.StreamSpliterators$WrappingSpliterator.lambda$initPartialTraversalState$0()Z (StreamSpliterators.java:294)
  at java.util.stream.StreamSpliterators$WrappingSpliterator$$Lambda$165.getAsBoolean()Z (Unknown Source)
  at java.util.stream.StreamSpliterators$AbstractWrappingSpliterator.fillBuffer()Z (StreamSpliterators.java:206)
  at java.util.stream.StreamSpliterators$AbstractWrappingSpliterator.doAdvance()Z (StreamSpliterators.java:169)
  at java.util.stream.StreamSpliterators$WrappingSpliterator.tryAdvance(Ljava/util/function/Consumer;)Z (StreamSpliterators.java:300)
  at java.util.Spliterators$1Adapter.hasNext()Z (Spliterators.java:681)
  at scala.collection.convert.Wrappers$JIteratorWrapper.hasNext()Z (Wrappers.scala:43)
  at geny.Generator$SelfClosing.generate(Lscala/Function1;)Lgeny/Generator$Action; (Generator.scala:228)
  at geny.Generator$Mapped.generate(Lscala/Function1;)Lgeny/Generator$Action; (Generator.scala:283)
  at geny.Generator.foreach(Lscala/Function1;)V (Generator.scala:49)
  at geny.Generator.foreach$(Lgeny/Generator;Lscala/Function1;)V (Generator.scala:49)
  at geny.Generator$Mapped.foreach(Lscala/Function1;)V (Generator.scala:281)
  at scala.meta.internal.mtags.OnDemandSymbolIndex.$anonfun$addSourceJar$2(Lscala/meta/internal/mtags/OnDemandSymbolIndex;Lscala/meta/io/AbsolutePath;)V (OnDemandSymbolIndex.scala:72)
  at scala.meta.internal.mtags.OnDemandSymbolIndex.$anonfun$addSourceJar$2$adapted(Lscala/meta/internal/mtags/OnDemandSymbolIndex;Lscala/meta/io/AbsolutePath;)Ljava/lang/Object; (OnDemandSymbolIndex.scala:71)
  at scala.meta.internal.mtags.OnDemandSymbolIndex$$Lambda$2743.apply(Ljava/lang/Object;)Ljava/lang/Object; (Unknown Source)
  at scala.meta.internal.io.PlatformFileIO$.withJarFileSystem(Lscala/meta/io/AbsolutePath;ZZLscala/Function1;)Ljava/lang/Object; (PlatformFileIO.scala:87)
  at scala.meta.internal.io.FileIO$.withJarFileSystem(Lscala/meta/io/AbsolutePath;ZZLscala/Function1;)Ljava/lang/Object; (FileIO.scala:44)
  at scala.meta.internal.mtags.OnDemandSymbolIndex.$anonfun$addSourceJar$1(Lscala/meta/internal/mtags/OnDemandSymbolIndex;Lscala/meta/io/AbsolutePath;)V (OnDemandSymbolIndex.scala:71)
  at scala.meta.internal.mtags.OnDemandSymbolIndex$$Lambda$2742.apply$mcV$sp()V (Unknown Source)
  at scala.meta.internal.mtags.OnDemandSymbolIndex.tryRun(Lscala/Function0;)V (OnDemandSymbolIndex.scala:125)
  at scala.meta.internal.mtags.OnDemandSymbolIndex.addSourceJar(Lscala/meta/io/AbsolutePath;)V (OnDemandSymbolIndex.scala:69)
  at scala.meta.internal.metals.MetalsLanguageServer.$anonfun$addSourceJarSymbols$1(Lscala/meta/internal/metals/MetalsLanguageServer;Lscala/meta/io/AbsolutePath;)Lscala/collection/concurrent/TrieMap; (MetalsLanguageServer.scala:1897)
  at scala.meta.internal.metals.MetalsLanguageServer$$Lambda$2735.apply()Ljava/lang/Object; (Unknown Source)
  at scala.meta.internal.mtags.OnDemandSymbolIndex.$anonfun$addSourceJarTopLevels$1(Lscala/meta/internal/mtags/OnDemandSymbolIndex;Lscala/meta/io/AbsolutePath;Lscala/Function0;)V (OnDemandSymbolIndex.scala:94)
  at scala.meta.internal.mtags.OnDemandSymbolIndex$$Lambda$2736.apply$mcV$sp()V (Unknown Source)
  at scala.meta.internal.mtags.OnDemandSymbolIndex.tryRun(Lscala/Function0;)V (OnDemandSymbolIndex.scala:125)
  at scala.meta.internal.mtags.OnDemandSymbolIndex.addSourceJarTopLevels(Lscala/meta/io/AbsolutePath;Lscala/Function0;)V (OnDemandSymbolIndex.scala:93)
  at scala.meta.internal.metals.MetalsLanguageServer.addSourceJarSymbols(Lscala/meta/io/AbsolutePath;)V (MetalsLanguageServer.scala:1886)
  at scala.meta.internal.metals.MetalsLanguageServer.$anonfun$indexDependencySources$5(Lscala/meta/internal/metals/MetalsLanguageServer;Ljava/util/HashSet;Lch/epfl/scala/bsp4j/DependencySourcesItem;Lscala/collection/mutable/HashSet;Ljava/lang/String;)V (MetalsLanguageServer.scala:1854)
  at scala.meta.internal.metals.MetalsLanguageServer.$anonfun$indexDependencySources$5$adapted(Lscala/meta/internal/metals/MetalsLanguageServer;Ljava/util/HashSet;Lch/epfl/scala/bsp4j/DependencySourcesItem;Lscala/collection/mutable/HashSet;Ljava/lang/String;)Ljava/lang/Object; (MetalsLanguageServer.scala:1845)
  at scala.meta.internal.metals.MetalsLanguageServer$$Lambda$2748.apply(Ljava/lang/Object;)Ljava/lang/Object; (Unknown Source)
  at scala.collection.TraversableLike$WithFilter.$anonfun$foreach$1(Lscala/collection/TraversableLike$WithFilter;Lscala/Function1;Ljava/lang/Object;)Ljava/lang/Object; (TraversableLike.scala:912)
  at scala.collection.TraversableLike$WithFilter$$Lambda$1673.apply(Ljava/lang/Object;)Ljava/lang/Object; (Unknown Source)
  at scala.collection.immutable.List.foreach(Lscala/Function1;)V (List.scala:392)
  at scala.collection.TraversableLike$WithFilter.foreach(Lscala/Function1;)V (TraversableLike.scala:911)
  at scala.meta.internal.metals.MetalsLanguageServer.$anonfun$indexDependencySources$2(Lscala/meta/internal/metals/MetalsLanguageServer;Ljava/util/HashSet;Lscala/collection/mutable/HashSet;Lch/epfl/scala/bsp4j/DependencySourcesItem;)V (MetalsLanguageServer.scala:1845)
  at scala.meta.internal.metals.MetalsLanguageServer.$anonfun$indexDependencySources$2$adapted(Lscala/meta/internal/metals/MetalsLanguageServer;Ljava/util/HashSet;Lscala/collection/mutable/HashSet;Lch/epfl/scala/bsp4j/DependencySourcesItem;)Ljava/lang/Object; (MetalsLanguageServer.scala:1844)
  at scala.meta.internal.metals.MetalsLanguageServer$$Lambda$2745.apply(Ljava/lang/Object;)Ljava/lang/Object; (Unknown Source)
  at scala.collection.Iterator.foreach(Lscala/Function1;)V (Iterator.scala:943)
  at scala.collection.Iterator.foreach$(Lscala/collection/Iterator;Lscala/Function1;)V (Iterator.scala:943)
  at scala.collection.AbstractIterator.foreach(Lscala/Function1;)V (Iterator.scala:1431)
  at scala.collection.IterableLike.foreach(Lscala/Function1;)V (IterableLike.scala:74)
  at scala.collection.IterableLike.foreach$(Lscala/collection/IterableLike;Lscala/Function1;)V (IterableLike.scala:73)
  at scala.collection.AbstractIterable.foreach(Lscala/Function1;)V (Iterable.scala:56)
  at scala.meta.internal.metals.MetalsLanguageServer.indexDependencySources(Lch/epfl/scala/bsp4j/DependencySourcesResult;)V (MetalsLanguageServer.scala:1844)
  at scala.meta.internal.metals.MetalsLanguageServer.$anonfun$indexWorkspace$9(Lscala/meta/internal/metals/MetalsLanguageServer;Lscala/meta/internal/metals/ImportedBuild;)V (MetalsLanguageServer.scala:1790)
  at scala.meta.internal.metals.MetalsLanguageServer$$Lambda$2727.apply$mcV$sp()V (Unknown Source)
  at scala.runtime.java8.JFunction0$mcV$sp.apply()Ljava/lang/Object; (JFunction0$mcV$sp.java:23)
  at scala.meta.internal.metals.MetalsLanguageServer.timedThunk(Ljava/lang/String;ZJLscala/Function0;)Ljava/lang/Object; (MetalsLanguageServer.scala:1666)
  at scala.meta.internal.metals.MetalsLanguageServer.indexWorkspace(Lscala/meta/internal/metals/ImportedBuild;)V (MetalsLanguageServer.scala:1790)
  at scala.meta.internal.metals.MetalsLanguageServer.$anonfun$connectToNewBuildServer$10(Lscala/meta/internal/metals/MetalsLanguageServer;Lscala/meta/internal/metals/ImportedBuild;)V (MetalsLanguageServer.scala:1582)
  at scala.meta.internal.metals.MetalsLanguageServer$$Lambda$2170.apply$mcV$sp()V (Unknown Source)
  at scala.meta.internal.metals.MetalsLanguageServer.$anonfun$profiledIndexWorkspace$2(Lscala/Function0;Lscala/Function0;)V (MetalsLanguageServer.scala:1694)
  at scala.meta.internal.metals.MetalsLanguageServer$$Lambda$2173.apply$mcV$sp()V (Unknown Source)
  at scala.runtime.java8.JFunction0$mcV$sp.apply()Ljava/lang/Object; (JFunction0$mcV$sp.java:23)
  at scala.meta.internal.metals.MetalsLanguageServer.timedThunk(Ljava/lang/String;ZJLscala/Function0;)Ljava/lang/Object; (MetalsLanguageServer.scala:1666)
  at scala.meta.internal.metals.MetalsLanguageServer.$anonfun$profiledIndexWorkspace$1(Lscala/meta/internal/metals/MetalsLanguageServer;Lscala/Function0;Lscala/Function0;)V (MetalsLanguageServer.scala:1694)
  at scala.meta.internal.metals.MetalsLanguageServer$$Lambda$2172.apply$mcV$sp()V (Unknown Source)
  at scala.runtime.java8.JFunction0$mcV$sp.apply()Ljava/lang/Object; (JFunction0$mcV$sp.java:23)
  at scala.concurrent.Future$.$anonfun$apply$1(Lscala/Function0;Lscala/runtime/BoxedUnit;)Ljava/lang/Object; (Future.scala:659)
  at scala.concurrent.Future$$$Lambda$116.apply(Ljava/lang/Object;)Ljava/lang/Object; (Unknown Source)
  at scala.util.Success.$anonfun$map$1(Lscala/util/Success;Lscala/Function1;)Ljava/lang/Object; (Try.scala:255)
  at scala.util.Success.map(Lscala/Function1;)Lscala/util/Try; (Try.scala:213)
  at scala.concurrent.Future.$anonfun$map$1(Lscala/Function1;Lscala/util/Try;)Lscala/util/Try; (Future.scala:292)
  at scala.concurrent.Future$$Lambda$117.apply(Ljava/lang/Object;)Ljava/lang/Object; (Unknown Source)
  at scala.concurrent.impl.Promise.liftedTree1$1(Lscala/Function1;Lscala/util/Try;)Lscala/util/Try; (Promise.scala:33)
  at scala.concurrent.impl.Promise.$anonfun$transform$1(Lscala/concurrent/impl/Promise$DefaultPromise;Lscala/Function1;Lscala/util/Try;)Lscala/concurrent/impl/Promise$DefaultPromise; (Promise.scala:33)
  at scala.concurrent.impl.Promise$$Lambda$118.apply(Ljava/lang/Object;)Ljava/lang/Object; (Unknown Source)
  at scala.concurrent.impl.CallbackRunnable.run()V (Promise.scala:64)
  at java.util.concurrent.ThreadPoolExecutor.runWorker(Ljava/util/concurrent/ThreadPoolExecutor$Worker;)V (ThreadPoolExecutor.java:1128)
  at java.util.concurrent.ThreadPoolExecutor$Worker.run()V (ThreadPoolExecutor.java:628)
  at java.lang.Thread.run()V (Thread.java:834)
```