@tgodzik 2020-05-15
Thanks for the contribution! I haven't seen the issue happen on my machine, but I might have just ignored some symptom. I will test it out.