@greg-a-atkinson 2020-05-15
> > Thanks for the contribution! I haven't seen the issue happen on my machine, but I might have just ignored some symptom. I will test it out.
> 
> @greg-a-atkinson I am unable to reproduce locally, there don't seem to be any recursive calls for that jar. Is it happening only on a particular system like on MacOS?

Hi @tgodzik, thanks for your reply.

The underlying system is an Ubuntu 18.04 docker container. Metals is running against OpenJDK 11.0.2 downloaded by jabba.

There was nothing in the metals log that indicated a problem. Indexing ran indefinitely while running the import build task. Running top showed 200-300% CPU on the metals JVM and an increasing heap, which ran out of memory in 15 to 30 minutes depending on the maximum heap size and the import build task didn't complete. I started by taking a heap dump of the metals JVM after it exhausted an 8GB heap. The heap dump showed a `java.nio.file.FileTreeWalker` consuming the entire heap with millions of `java.nio.file.FileTreeWalker$DirectoryNode` instances and the stack trace included above confirmed the code path leading to the issue. I then added logging to metals to isolate the code path and the problematic JAR and was able to confirm the list of directories above that were repeatedly walked.

I too am surprised by this issue, it must be an edge case because I'm sure many other developers are using VS Code with Metals successfully, but I had the same problem attempting to import the Metals and FS2 code bases. It does appear to be something specific to the versions being used. I suspect the OpenJDK version because it looks like file tree walker implementation is JVM implementation specific.

Any ideas?

Thanks