@greg-a-atkinson 2020-05-15
@tgodzik , see below an extract from a 2GB heap dump for this issue: 7 million `java.nio.file.FileTreeWalker$DirectoryNode` instances for the same directories from the problematic JAR:

![image](https://user-images.githubusercontent.com/11074540/82055611-a97e5880-96b8-11ea-808f-5ae2c65138bc.png)