@greg-a-atkinson 2020-05-15
@tgodzik - found it. It is OpenJDK version 11.0.2 that is the problem. This was fixed in OpenJDK 11.0.3:

S8197398: (zipfs) Files.walkFileTree walk indefinitelly while processing JAR file with "/" as a directory inside.

https://adoptopenjdk.net/release_notes.html